src/requirements.txt
python-telegram-bot==21.6
psycopg2-binary
requests
httpx
reportlab